docker_switch: docker-current <-> docker-latest
This pair of sub-subtests simply switches back and forth between
docker-current and docker-latest (in RHEL and CentOS).  Then after
restarting the docker daemon, it confirms the version change. This test
is intended for use along with a customized sequence of other tests.
There is no provison made for recovering the system if the one of these
tests fail.  Therefor, both tests are disabled/excluded by default.

Signed-off-by: Chris Evich <cevich@redhat.com>