Workaround RH BZ 1393572 & docker_cli/negativeusage/iv4
This sub-subtest is failing on all platforms and I see little hope for
the UX-related issue being addressed anytime soon (IMHO).  Workaround
the problem for now, so the sub-subtest can be used to possible detect
other problems until bug is fixed.  Add a config. comment that explains
this also and preserves the reproducer-configuration.

Signed-off-by: Chris Evich <cevich@redhat.com>