Add class to represent canonical distro. names
It's desireable to have a single, simple set of
distribution names to act upon.  However,
occasionally the detection methods need to change.
Add a class that behaves like a regular string,
but the allows detection-methods to be added/changed
over time to support new and/or overlaping types
of detection.

In additon, because it acts like a normal string,
any of the instance methods can be run against any
other pre-existing string from any source.  This
supports use of those methods without forcing instance
creation to always use the detection methods.

Signed-off-by: Chris Evich <cevich@redhat.com>

Rename variables for readability

Signed-off-by: Ed Santiago <santiago@redhat.com>